refactor: extract tooltip position logic to separate mixin
